refactor: integration tests
